Adicionando token e autenticação para segurança de rotas

+++ bd-projeto-guiado-III/gameStore/src/controller/consolesController.js
@@ -16,6 +16,7 @@ const findAllConsoles = async (req, res) => {
       if (erro) {
         return res.status(403).send('Acesso não autorizado!')
       }
+
       const allConsoles = await ConsolesModel.find();
       res.status(200).json(allConsoles);
     })
@@ -27,8 +28,23 @@ const findAllConsoles = async (req, res) => {
 
 const findConsoleById = async (req, res) => {
   try {
-    const findConsole = await ConsolesModel.findById(req.params.id);
-    res.status(200).json(findConsole);
+    const authHeader = req.get('authorization')
+
+    if (!authHeader) {
+      return res.status(401).send('Você esqueceu de passar as informações de autorização!')
+    }
+
+    const token = authHeader.split(' ')[1]
+
+    jwt.verify(token, SECRET, async function (erro) {
+      if (erro) {
+        return res.status(403).send('Acesso não autorizado!')
+      }
+
+      const findConsole = await ConsolesModel.findById(req.params.id);
+      res.status(200).json(findConsole);
+    })
+
   } catch (error) {
     console.error(error);
     res.status(500).json({ message: error.message });
@@ -37,30 +53,43 @@ const findConsoleById = async (req, res) => {
 
 const addNewConsole = async (req, res) => {
   try {
-    const {
-      name,
-      developer,
-      releaseDate,
-      display,
-      storageCapacities,
-      numberOfPlayers,
-      available,
-      description,
-    } = req.body;
-    const newConsole = new ConsolesModel({
-      name,
-      developer,
-      releaseDate,
-      display,
-      storageCapacities,
-      numberOfPlayers,
-      available,
-      description,
-    });
+    const authHeader = req.get('authorization')
 
-    const savedConsole = await newConsole.save();
+    if (!authHeader) {
+      return res.status(401).send('Você esqueceu de passar as informações de autorização!')
+    }
 
-    res.status(201).json({ message: "New console successfully added", savedConsole });
+    const token = authHeader.split(' ')[1]
+
+    jwt.verify(token, SECRET, async function (erro) {
+      if (erro) {
+        return res.status(403).send('Acesso não autorizado!')
+      }
+      const {
+        name,
+        developer,
+        releaseDate,
+        display,
+        storageCapacities,
+        numberOfPlayers,
+        available,
+        description,
+      } = req.body;
+      const newConsole = new ConsolesModel({
+        name,
+        developer,
+        releaseDate,
+        display,
+        storageCapacities,
+        numberOfPlayers,
+        available,
+        description,
+      });
+  
+      const savedConsole = await newConsole.save();
+  
+      res.status(201).json({ message: "New console successfully added", savedConsole });
+    })   
   } catch (error) {
     console.error(error);
     res.status(500).json(error.message);
@@ -69,6 +98,44 @@ const addNewConsole = async (req, res) => {
 
 const updateConsole = async (req, res) => {
   try {
+
+    const authHeader = req.get('authorization')
+
+    if (!authHeader) {
+      return res.status(401).send('Você esqueceu de passar as informações de autorização!')
+    }
+
+    const token = authHeader.split(' ')[1]
+
+    jwt.verify(token, SECRET, async function (erro) {
+      if (erro) {
+        return res.status(403).send('Acesso não autorizado!')
+      }
+
+      const {
+        name,
+        developer,
+        releaseDate,
+        display,
+        storageCapacities,
+        numberOfPlayers,
+        available,
+        description,
+      } = req.body;
+      const updateConsole = await ConsolesModel.findByIdAndUpdate(req.params.id, {
+        name,
+        developer,
+        releaseDate,
+        display,
+        storageCapacities,
+        numberOfPlayers,
+        available,
+        description,
+      });
+  
+      res.status(200).json({ message: "Console successfully updated", updateConsole });
+    })
+
     const {
       name,
       developer,
@@ -99,10 +166,24 @@ const updateConsole = async (req, res) => {
 
 const deleteConsole = async (req, res) => {
   try {
-    const { id } = req.params;
-    const deleteConsole = await ConsolesModel.findByIdAndDelete(id);
-    const message = `Console with id ${deleteConsole.name} was successfully deleted`;
-    res.status(200).json({ message });
+
+    const authHeader = req.get('authorization')
+
+    if (!authHeader) {
+      return res.status(401).send('Você esqueceu de passar as informações de autorização!')
+    }
+
+    const token = authHeader.split(' ')[1]
+
+    jwt.verify(token, SECRET, async function (erro) {
+      if (erro) {
+        return res.status(403).send('Acesso não autorizado!')
+      }
+      const { id } = req.params;
+      const deleteConsole = await ConsolesModel.findByIdAndDelete(id);
+      const message = `Console with id ${deleteConsole.name} was successfully deleted`;
+      res.status(200).json({ message });
+    })      
   } catch (error){
     console.error(error);
     res.status(500).json({ message: error.message });
